+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 * Added: Stop launch with different status than success
+* Added: Add the option to log on a launch
 
 ## 1.0.0 - 2020-01-24