+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 1.2.0 - 2020-02-20
 
 * Added: Launch, Suite, Test and Step blocks for the report portal DSL
 * Added: Option to invoke generic requests with Invoke-RPRequest